Rename schemas to models, rename service files to their service name
